Update README.md
The private key of electronic elections is a very critical piece of information that, with an incorrect or improper use, may disrupt
the elections results. To enforce the privacy and security of the private key, secret sharing schemes (or threshold schemes) are used to generate a distributed key into several entities. In this fashion, a threshold of
at least t out of the n entities will be necessary to decrypt votes. We study in this work the feasibility of developing ElGamal cryptosystem
and Shamir’s secret sharing scheme into JavaCards, whose API gives no support for it.